pyproject.toml
[tool.poetry]
name = "nommo-engine"
version = "0.1.0"
description = "Scientific life emergence simulator with real physics and thermodynamics"
authors = ["Rico Ledan"]
readme = "README.md"
packages = [{include = "nommo"}]
license = "MIT"
homepage = "https://github.com/ricoledan/nommo-engine"
repository = "https://github.com/ricoledan/nommo-engine"
keywords = ["simulation", "emergence", "physics", "chemistry", "molecular-dynamics", "artificial-life"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Chemistry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.0"
scipy = "^1.11.0"
matplotlib = "^3.8.0"
click = "^8.1.0"
rich = "^13.7.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.0"
h5py = "^3.10.0"
numba = "^0.58.0"
joblib = "^1.3.0"
networkx = "^3.2.0"
pandas = "^2.1.0"
plotly = "^5.18.0"
tqdm = "^4.66.0"
watchdog = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.5.0"
pytest-benchmark = "^4.0.0"
hypothesis = "^6.92.0"
black = "^23.12.0"
ruff = "^0.1.0"
mypy = "^1.7.0"
ipython = "^8.18.0"
jupyter = "^1.0.0"
notebook = "^7.0.0"
pre-commit = "^3.6.0"
mkdocs = "^1.5.0"
mkdocs-material = "^9.5.0"
mkdocstrings = "^0.24.0"
mkdocstrings-python = "^1.7.0"

[tool.poetry.group.viz]
optional = true

[tool.poetry.group.viz.dependencies]
vispy = "^0.14.0"
vpython = "^7.6.0"
pyopengl = "^3.1.0"

[tool.poetry.scripts]
nommo = "nommo.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = ["E501", "B008", "C901", "B904"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=nommo --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["nommo"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]